Pull from local manifest if stored
There's still a need to fetch the remote manifests,
since I'm not indexing these by their names. If we
decide we want to force users into always using the same
name (e.g. docker.io/busybox:latest vs busybox), then
we can check to see if a manifest exists and not pull it.
Or there might be another solution I'm not thinking of.

Signed-off-by: Christy Perez <christy@linux.vnet.ibm.com>